requirements.txt
pandas
numpy
matplotlib
scikit-learn
onnxruntime
skl2onnx
requests
python-dotenv
